+++ core/src/inspect/mod.rs
@@ -1,9 +1,11 @@
 mod column;
+mod extension;
 mod relation;
 mod schema;
 mod types;
 
 pub use column::*;
+pub use extension::*;
 pub use relation::*;
 pub use schema::*;
 pub use types::*;